requirements.txt
python-usbtmc~=0.8
pyserial~=3.5
pyusb~=1.2.1
python-dotenv~=1.0.1